refactor(web): 공통 로그인 feature 분리
