+++ apps/web/src/features/login/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginBanner } from "./LoginBanner";
+export { default as LoginButton } from "./LoginButton";
